Overhaul Java load balancer
This patch is an overhaul of the Java load balancer.

The main goal was to add support for v1 of the AWS SDK for Java, whereas
the existing code only support v2.

README.md was also completely rewritten, to explain the two SDK versions
and how to use them.

The v1 support is very different, and much simpler, than the old v2
implementation. The reason why the simpler implementation was possible
is a different (and better) understanding of how SSL would be used
in a load-balanced Alternator setup. The README.md suggests how we
should simplify the v2 implementation to be like the new v1
implementation, but actually doing this is left to the future.
We may want to do similar changes also to the implementations in
other languages.

Additionally, the pom.xml was cleaned from unnecessary stuff, and
building of the library and demos actually works now. The demo is
built, but no longer part of the generated jar file - which only
contains the minimum needed for applications - these of course don't
need the demo.

The existing code also had errors dealing with various combinations
of SSL and authentication checks, which this patch fixes.

In this patch, I replaced the separate GetItem/PutItem/CreateTable
demos by one demo for each version (Demo1.java for v1, Demo2.java for v2)
which does a bunch of requests. It makes it easy to demonstrate how
different requests in sequence contact different Scylla nodes.
The demos enable detailed logging to show what the load balancer
is choosing - these log messages are invisible by default.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>