Add cpp load balancer
The load balancer is based on a simple wrapper over DynamoDBClient:
AlternatorClient. The wrapper needs to be constructed with the following
information:
 - used protocol (http/https)
 - address of one of the alternator nodes
 - used port

AlternatorClient is able to fetch information about live nodes
from the alternator cluster. It's performed once during itsa creation,
but can also be done continuously in a separate thread by invoking
the StartNodeUpdater() helper method.

Fixes #4

Message-Id: <beb88540e3b292f076c5275950f5165d4315ab29.1611573321.git.sarna@scylladb.com>