Remove excessive printouts
This patch removes excessive printouts on the Go, Javascript and Python
drivers. While printing every decision on every request is pretty for
demonstrating the load balancing, it is annoying when used in production
and can needlessly fill the logs.

As far as I can tell, the Java driver did not have such excessive
printouts (it has logs in the FINE level, but those will be normally
not visible).

Signed-off-by: Nadav Har'El <nyh@scylladb.com>

Closes #19