dns: add example DNS server which returns all nodes instead of one
This patch contains a proof-of-concept of the "DNS Round Robin" idea -
where instead of the DNS returning a single random live Scylla node (as
DynamoDB does, and our existing dns-loadbalancer.py POC does), the new
code dns/dns-loadbalancer-rr.py returns **all** the live nodes for
each DNS request.

To support old resolvers that resolve a name to the first returned DNS
response, the POC server returns a random cyclical rotation of the list of
live nodes - so that the first one will be different each time. But our
hope is that modern resolvers and libraries and SDKs will actually be able
to make use of the complete list. Specifically, that even if DNS responses
are cached for the full TTL (5 seconds), then during this duration we hope
that different connections, different processes or different client
machines will choose different Scylla nodes and not the same one.

Issue #105 is about verifying that this "hope" actually happens, and
checking whether dns-loadbalancer-rr.py really has any advantages compared
to dns-loadbalancer.

Fixes #14

Signed-off-by: Nadav Har'El <nyh@scylladb.com>